Berry suppress warnings (#18626)

+++ lib/libesp32/berry/src/be_introspectlib.c
@@ -109,7 +109,7 @@ static int m_toptr(bvm *vm)
             be_pushcomptr(vm, var_toobj(v));
             be_return(vm);
         } else if (var_type(v) == BE_INT) {
-            be_pushcomptr(vm, (void*) var_toint(v));
+            be_pushcomptr(vm, (void*) (intptr_t) var_toint(v));
             be_return(vm);
         } else {
             be_raise(vm, "value_error", "unsupported for this type");
@@ -126,7 +126,7 @@ static int m_fromptr(bvm *vm)
         if (be_iscomptr(vm, 1)) {
             v = be_tocomptr(vm, 1);
         } else {
-            v = (void*) be_toint(vm, 1);
+            v = (void*) (intptr_t) be_toint(vm, 1);
         }
         if (v) {
             bgcobject *ptr = (bgcobject*)v;
